Extract ConfigBuilder from Config class
