+++ lib/tapioca.rb
@@ -28,13 +28,14 @@ rescue
   nil
 end
 
-require_relative "tapioca/loader"
-require_relative "tapioca/constant_locator"
-require_relative "tapioca/config"
-require_relative "tapioca/generator"
-require_relative "tapioca/cli"
-require_relative "tapioca/gemfile"
-require_relative "tapioca/compilers/symbol_table_compiler"
-require_relative "tapioca/compilers/symbol_table/symbol_generator"
-require_relative "tapioca/compilers/symbol_table/symbol_loader"
-require_relative "tapioca/version"
+require "tapioca/loader"
+require "tapioca/constant_locator"
+require "tapioca/config"
+require "tapioca/config_builder"
+require "tapioca/generator"
+require "tapioca/cli"
+require "tapioca/gemfile"
+require "tapioca/compilers/symbol_table_compiler"
+require "tapioca/compilers/symbol_table/symbol_generator"
+require "tapioca/compilers/symbol_table/symbol_loader"
+require "tapioca/version"
